Module aws_sdk_lexruntimev2::model

Modules

active_context

See ActiveContext

active_context_time_to_live

See ActiveContextTimeToLive

button

See Button

confidence_score

See ConfidenceScore

dialog_action

See DialogAction

image_response_card

See ImageResponseCard

intent

See Intent

interpretation

See Interpretation

message

See Message

sentiment_response

See SentimentResponse

sentiment_score

See SentimentScore

session_state

See SessionState

slot

See Slot

value

See Value

Structs

ActiveContext

Contains information about the contexts that a user is using in a

ActiveContextTimeToLive

The time that a context is active. You can specify the time to live

Button

A button that appears on a response card show to the user.

ConfidenceScore

Provides a score that indicates the confidence that Amazon Lex V2 has that

DialogAction

The next action that Amazon Lex V2 should take.

ImageResponseCard

A card that is shown to the user by a messaging platform. You define

Intent

The current intent that Amazon Lex V2 is attempting to fulfill.

Interpretation

An intent that Amazon Lex V2 determined might satisfy the user's utterance.

Message

Container for text that is returned to the customer..

SentimentResponse

Provides information about the sentiment expressed in a user's

SentimentScore

The individual sentiment responses for the utterance.

SessionState

The state of the user's session with Amazon Lex V2.

Slot

A value that Amazon Lex V2 uses to fulfill an intent.

Value

The value of a slot.

Enums

ConfirmationState

DialogActionType

IntentState

MessageContentType

SentimentType

Shape