Struct aws_sdk_lexruntimev2::output::recognize_utterance_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for RecognizeUtteranceOutput

Implementations

impl Builder

pub fn input_mode(self, input: impl Into<String>) -> Self

Indicates whether the input mode to the operation was text or speech.

pub fn set_input_mode(self, input: Option<String>) -> Self

pub fn content_type(self, input: impl Into<String>) -> Self

Content type as specified in the responseContentType in the request.

pub fn set_content_type(self, input: Option<String>) -> Self

pub fn messages(self, input: impl Into<String>) -> Self

A list of messages that were last sent to the user. The messages are ordered based on the order that you returned the messages from your Lambda function or the order that the messages are defined in the bot.

The messages field is compressed with gzip and then base64 encoded. Before you can use the contents of the field, you must decode and decompress the contents. See the example for a simple function to decode and decompress the contents.

pub fn set_messages(self, input: Option<String>) -> Self

pub fn interpretations(self, input: impl Into<String>) -> Self

A list of intents that Amazon Lex V2 determined might satisfy the user's utterance.

Each interpretation includes the intent, a score that indicates how confident Amazon Lex V2 is that the interpretation is the correct one, and an optional sentiment response that indicates the sentiment expressed in the utterance.

The interpretations field is compressed with gzip and then base64 encoded. Before you can use the contents of the field, you must decode and decompress the contents. See the example for a simple function to decode and decompress the contents.

pub fn set_interpretations(self, input: Option<String>) -> Self

pub fn session_state(self, input: impl Into<String>) -> Self

Represents the current state of the dialog between the user and the bot.

Use this to determine the progress of the conversation and what the next action might be.

The sessionState field is compressed with gzip and then base64 encoded. Before you can use the contents of the field, you must decode and decompress the contents. See the example for a simple function to decode and decompress the contents.

pub fn set_session_state(self, input: Option<String>) -> Self

pub fn request_attributes(self, input: impl Into<String>) -> Self

The attributes sent in the request.

The requestAttributes field is compressed with gzip and then base64 encoded. Before you can use the contents of the field, you must decode and decompress the contents.

pub fn set_request_attributes(self, input: Option<String>) -> Self

pub fn session_id(self, input: impl Into<String>) -> Self

The identifier of the session in use.

pub fn set_session_id(self, input: Option<String>) -> Self

pub fn input_transcript(self, input: impl Into<String>) -> Self

The text used to process the request.

If the input was an audio stream, the inputTranscript field contains the text extracted from the audio stream. This is the text that is actually processed to recognize intents and slot values. You can use this information to determine if Amazon Lex V2 is correctly processing the audio that you send.

The inputTranscript field is compressed with gzip and then base64 encoded. Before you can use the contents of the field, you must decode and decompress the contents. See the example for a simple function to decode and decompress the contents.

pub fn set_input_transcript(self, input: Option<String>) -> Self

pub fn audio_stream(self, input: ByteStream) -> Self

The prompt or statement to send to the user. This is based on the bot configuration and context. For example, if Amazon Lex V2 did not understand the user intent, it sends the clarificationPrompt configured for the bot. If the intent requires confirmation before taking the fulfillment action, it sends the confirmationPrompt. Another example: Suppose that the Lambda function successfully fulfilled the intent, and sent a message to convey to the user. Then Amazon Lex V2 sends that message in the response.

pub fn set_audio_stream(self, input: Option<ByteStream>) -> Self

pub fn build(self) -> RecognizeUtteranceOutput

Consumes the builder and constructs a RecognizeUtteranceOutput